Resources Front End Pages
Restored: main.dart
Created: resources.dart, categoryName.dart
Worked on: general layout on Resources page
